requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
boto3==1.34.0
django-storages==1.14.2
Pillow==10.1.0
ffmpeg-python==0.2.0
django-cors-headers==4.3.1
python-decouple==3.8
